Add tools for calculating and caching distances.
Loop and sidechain RMSDs are important for assessing model quality at
many stages of the design pipeline.  Loop RMSD is pretty easy to work
with, but sidechain RMSD is less well defined.  The problem is that
different designs are interested in different sidechains.  However, this
information is encoded in the restraint file associated with each
design.  The distance module added here parses the restraint file to
determine which distances are relevant to the design.