Set the kr_build score function to talaris2013_cst.
It's probably a good thing to be explicit about which score function is
being used.  But this commit was motivated by the fact that rosetta
scripts was using the wrong score function (vanilla talaris2013) to
rescore the pose at the end of each run.  This was causing the scores
outputted in the pdb file to disagree with the scores reported by
LoopProtocol.

This solution is still a little bit of a hack, because LoopProtocol
isn't necessarily using talaris2013_cst.  This is why I also removed the
constraints:cst:fa_weights flag; it could have caused the constraint
weights used by LoopProtocol to disagree with the reported weights.  So
it's still possible to confuse things, but I think the kr_build script
should be robust to the usual level of tweaking.