Make minor improvements to lots of things.
I added the ability for workspaces to look in several places for
required files.  This should make it easier to use different settings
for different phases of the design.  I also added loopmodel and fixbb
rosetta scripts to the workspaces, so that they don't have to be so
one-size-fits-all.  I also did some general refactoring.