Automatically create workspaces for long jobs.
This was an fairly straight-forward application of being able to create
workspaces from directories.  Before, the user would have to specify a
"workspace factory" to big_job.initiate().  It was not clear how this
system would deal with "round 2" style workspaces.  Now, initiate()
takes no arguments and always constructs the right workspace on its own.