Get rid of workspaces for the "picking" steps.
Previously there was a workspace for each step in the pipeline that
filtered results from the step just before.  I came to realize that this
was unnecessary.  The filter workspaces basically just had two
subdirectories -- inputs and outputs -- which were both just filled with
symlinks.  I was able to cut out a lot of unnecessary symlinks by
getting rid of these directories.

This commit also changes the Workspace class hierarchy.  There is now a
BigJobWorkspace subclass of Workspace.  BigJobWorkspace should be
inherited by workspace that use the cluster to generate lots of data.