Add 07_validate_designs.py and rename some files.
This commit adds the validation script.  While I was doing this, I
realized that it was not sufficient to have a single loopmodel XML
script, because I wanted to use different algorithms for the model
building and validation steps.  I added a second loopmodel script and
renamed all the existing XML scripts using more consistent names.