Make structures.load() more user-friendly.
The logic for deciding whether or not to use workspaces is now inside
structures.load(), which means it doesn't have to be replicated in every
script that calls it.  This also fixed bugs in fetch_data and push_data
that were causing the --remote argument to be ignored in some cases.
This commit also makes some small changes to the exception hierarchy.