Add fetch_and_cache_models.py
This script just waits for data generated by big cluster jobs, downloads
it, then caches it.  It's a nice way to have my cake and eat it too: I
don't have to wait to download and cache everything at the end, which
takes a long time, and I don't have to do any caching in the cluster
jobs themselves, which is a prospect rife with race-conditions.