Make sure all the rosetta paths work out.
I realized that you could confuse the code pretty easily by giving it
the wrong path to rosetta.  First, in most checkouts it expects the path
to rosetta/main, which is counter-intuitive.  Second, if you give it the
wrong path it won't crash until you submit a job, in which case you'll
get a cryptic message to stderr.  I tried to be more clear about what
path it wants, and to add better error checking in the submission
scripts themselves.

There is still one problem: PIP expects to find the rosetta_scripts
binary here:

    <workspace>/rosetta/source/bin/rosetta_scripts

In my experience, this file gets created when you compile everything.
However, for Marco, all the compiled binaries had suffixes like
'.linuxgccrelease'.  I'm not sure why this happened, and I don't know
how to deal with it in the general case.