Make a special command to link to remote designs.
This command is available as an option to the 01_setup script.  To
support this change, I moved all of the rsync code out of the scripts
directory and into the pipeline module.  The setup script will now only
ask for an rsync path if you're setting up a remote link, which is nice
because it's otherwise a difficult option to explain.