Dump the bizarre .gitignore strategy I'd been using.
Previously everything was ignored by default and only certain files were
specifically un-ignored.  This frequently caused me to forget to add new
files to commits.  The purpose of this strategy was to keep me from
checking-in project specific files when I was running PIP out of this
directory, but I don't do that anymore.