Install PIP system-wide using setuptools.
Before, you would use pip by cloning the repository and (typically)
making workspaces right there (so the commands would be easy to run).
However, this approach led to a proliferation of pip repositories and a
mixing of workspaces from different projects.  It was also quite non-
standard.

I decided to address these issues by making a single ``pull_into_place``
command that would be installed system-wide using ``setuptools``.  All
of the scripts that used to make up PIP are now available as sub-
commands of ``pull_into_place``.  It is even possible for sub-commands
to be installed from third-party packages.  Dependencies are handled in
a fairly sophisticated way, so that the large number of packages
required by the analysis scripts are not required on the cluster, but
can still be easily installed.