Interpret input paths relative to the workspace.
Previously jobs would be given paths to their inputs relative to the
directory the job was launched from.  However, this broke when I made
jobs run from the workspace directory.  Now the jobs are only given the
names of the input files (not the paths) and they assemble the paths
themselves.