Use Pareto fronts to pick designs to validate.
Previously I picked designs by boltzmann-weighting the total score term.
But this approach was deficient for two reasons:

1. It didn't use information from any of the filters Cody added.
2. It was meant for when almost every design (i.e. 50 of 100,000) would
   be filtered out (now I'm keeping 500 of 20,000).

I stumbled on the idea of validating the design of the Pareto front of
certain metrics.  I really like this idea because it yields a good
diversity of designs, since designs that score well for different
reasons tend to have different sequences.  (In fact, the problem I was
originally trying to solve with the Boltzmann-weighting was that designs
that score well for the same reason tend to have similar sequences.)