Refactor how designs are picked for validation.
The new process requires the user to provide a file specifying which
metrics to set thresholds for and which to include in the Pareto front.
The advantage of this approach is that this file is much more expressive
than the command-line, and also serves as a paper-trail of what actually
happened in the pipeline.